refactor: 프로젝트 초대 조회 DTO로 보냄
