+++ src/main/java/com/pm/projectmanager/domain/notification/NotificationService.java
@@ -2,10 +2,12 @@ package com.pm.projectmanager.domain.notification;
 
 import java.util.List;
 import java.util.Set;
+import java.util.stream.Collectors;
 
 import org.springframework.stereotype.Service;
 
 import com.pm.projectmanager.common.RedisService;
+import com.pm.projectmanager.domain.notification.dto.InviteResponseDto;
 import com.pm.projectmanager.security.UserDetailsImpl;
 
 import lombok.RequiredArgsConstructor;
@@ -16,7 +18,11 @@ public class NotificationService {
 
 	private final RedisService redisService;
 
-	public List<String> getInvites(UserDetailsImpl userDetails) {
-		return redisService.getInvites(userDetails.getUsername());
+	public List<InviteResponseDto> getInvites(UserDetailsImpl userDetails) {
+
+		return redisService.getInvites(userDetails.getUsername())
+			.stream()
+			.map(InviteResponseDto::new)
+			.collect(Collectors.toList());
 	}
 }
