+++ src/main/java/com/pm/projectmanager/domain/notification/dto/InviteResponseDto.java
@@ -1,0 +1,15 @@
+package com.pm.projectmanager.domain.notification.dto;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class InviteResponseDto {
+	private String projectId;
+
+	public InviteResponseDto(String projectId) {
+		this.projectId = projectId;
+	}
+
+}
