+++ src/main/java/com/pm/projectmanager/domain/user/dto/UpdateRequestDto.java
@@ -1,4 +1,4 @@
-package com.pm.projectmanager.domain.dto;
+package com.pm.projectmanager.domain.user.dto;
 
 import lombok.Getter;
